only loop over class callbacks that were valid for the initial data
this could cause issues in cases where a class callback returned an
instance of an unrelated class, where that class also has a class
callback. when looping over the callbacks, it would be dependent on hash
ordering whether the new class was tried first (in which case it would
be ignored) or tried last (in which case it would be run). now, it
collects the class callbacks to run first, and then dies if it comes
across a replacement object that doesn't work for the callbacks that
were chosen.